+++ lib/test_prof/event_prof/custom_events/sidekiq_jobs.rb
@@ -1,28 +1,5 @@
 # frozen_string_literal: true
 
-module TestProf::EventProf::CustomEvents
-  module SidekiqJobs # :nodoc: all
-    module ClientPatch
-      def raw_push(*)
-        return super unless Sidekiq::Testing.inline?
-        SidekiqJobs.track { super }
-      end
-    end
-
-    class << self
-      def setup!
-        Sidekiq::Client.prepend ClientPatch
-      end
-
-      def track
-        ActiveSupport::Notifications.instrument(
-          "sidekiq.jobs"
-        ) { yield }
-      end
-    end
-  end
-end
-
 TestProf::EventProf::CustomEvents.register("sidekiq.jobs") do
   if TestProf.require(
     "sidekiq/testing",
@@ -32,7 +9,12 @@ TestProf::EventProf::CustomEvents.register("sidekiq.jobs") do
       Make sure that "sidekiq" gem is in your Gemfile.
     MSG
   )
-    TestProf::EventProf::CustomEvents::SidekiqJobs.setup!
+    TestProf::EventProf.monitor(
+      Sidekiq::Client,
+      "sidekiq.jobs",
+      :raw_push,
+      guard: ->(*) { Sidekiq::Testing.inline? }
+    )
     TestProf::EventProf.configure do |config|
       config.rank_by = :count
     end