+++ lib/test_prof/factory_prof/factory_builders/factory_bot.rb
@@ -2,6 +2,7 @@
 
 require "test_prof/factory_prof/factory_bot_patch"
 require "test_prof/factory_bot"
+require "test_prof/ext/factory_bot_strategy"
 
 module TestProf
   module FactoryProf
@@ -9,23 +10,7 @@ module TestProf
       # implementation of #patch and #track methods
       # to provide unified interface for all factory-building gems
       class FactoryBot
-        # FactoryBot 5.0 uses strategy classes for associations,
-        # older versions and top-level invocations use Symbols
-        using(Module.new do
-          refine Symbol do
-            def create?
-              self == :create
-            end
-          end
-
-          if defined?(::FactoryBot::Strategy::Create)
-            refine Class do
-              def create?
-                self <= ::FactoryBot::Strategy::Create
-              end
-            end
-          end
-        end)
+        using TestProf::FactoryBotStrategy
 
         # Monkey-patch FactoryBot / FactoryGirl
         def self.patch